requirements.txt
tensorflow
numpy
Pillow
matplotlib
python-osc
numpy




